[GraphQL] Refactor cache generation to be generic.
